requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==0.5.0
Django==3.0
django-heroku==0.3.1
django-leaflet==0.28.2
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
python-decouple==3.6
pytz==2021.3
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
unicodecsv==0.13.0
urllib3==1.26.8
whitenoise==6.2.0
